moved stuff around and added teamcity script
