+++ backup/barman_metrics/bin/montagu_metrics/requirements.txt
@@ -1,0 +1,1 @@
+flask>=0.12.4
