+++ backup/barman_metrics/bin/requirements-dev.txt
@@ -1,0 +1,2 @@
+pytest>=3.5.1
+teamcity-messages>=1.21
