+++ backup/barman_metrics/tests/test_main.py
@@ -1,4 +1,4 @@
-from metrics.main import parse_status, metrics
+from bin.main import parse_status, metrics
 
 fake_barman_status = """
     Active: True
