--- backup/metrics/requirements-dev.txt
@@ -1,1 +1,0 @@
-pytest>=3.5.1
